requirements.txt
Flask==2.3.3
pandas==2.1.0
python-dateutil==2.8.2
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
# Toxicity detection dependencies
google-api-python-client>=2.0.0
tqdm>=4.65.0
# OpenAI moderation dependencies
openai>=1.0.0
# Text redaction dependencies
opencv-python>=4.8.0
pytesseract>=0.3.10
Pillow>=10.0.0
